pyproject.toml
[tool.poetry]
name = "PALM"
version = "0.1.0"
description = "Protein Aggregation prediction with Language Model embeddings"
authors = ["Ethan Eschbach, sbxw <sbxw@novonordisk.com>, dpkn <dpkn@novonordisk.com>, kcsl <kcsl@novonordisk.com>, hspu <hspu@novonordisk.com>"]
readme = "README.md"
packages = [{ include = "src", from = "." }]

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
# Core ML dependencies from MLflow conda.yaml
mlflow = "^2.13.0"
torch = "^2.3.0"
transformers = "^4.56.2"
lightning = "2.4.0"
numpy = "^1.26.4"
pandas = "^2.2.2"
scipy = "^1.13.1"
scikit-learn = "^1.3.2"

# Visualization and utilities
matplotlib = "^3.9.0"
seaborn = "^0.13.2"
rich = "^13.7.1"
tqdm = "^4.66.4"
statsmodels = "^0.14.5"

# Configuration and data
hydra-core = "^1.3.2"
pyyaml = "^6.0.1"
pydantic = "^2.3.0"

# Additional ML/NLP tools
sentencepiece = "^0.1.99"
biopython = "^1.83"  # Added for FASTA file handling

# Infrastructure and API
click = "^8.1.7"
requests = "^2.32.3"
boto3 = "^1.40.36"
cloudpickle = "^3.0.0"
dill = "^0.3.8"
joblib = "^1.3.2"

# Type checking and validation
mypy = "^1.7.1"
attrs = "^25.3.0"

# Other dependencies
astunparse = "^1.6.3"
cffi = "^1.16.0"
defusedxml = "^0.7.1"
packaging = "^24.1"
psutil = "^5.9.8"
levenshtein = "^0.23.0"
regex = "2025.9.1"
skops = "^0.12.0"
rdkit = "^2023.9.2"
StrEnum = "^0.4.15"
omegaconf = "^2.3.0"  # Added - required for Hydra configs

# Platform-specific tensorflow (optional - remove if not needed)
tensorflow = [
    { version = "2.15.0", platform = "linux", markers = "platform_machine!='arm64' and platform_machine!='aarch64'" },
    { version = "2.15.0", platform = "darwin", markers = "platform_machine!='arm64' and platform_machine!='aarch64'" },
]
tensorflow-macos = { version = "2.15.0", platform = "darwin", markers = "platform_machine=='arm64'" }
tensorflow-io-gcs-filesystem = "0.34.0"

[tool.poetry.group.dev.dependencies]
black = "^23.11.0"
isort = "^5.12.0"
pre-commit = "^3.3.3"
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# Ruff configuration
[tool.ruff]
select = ["E", "F", "I", "N", "C90", "B", "PL"]
ignore-init-module-imports = true
fixable = ["F401", "I"]
ignore = ["E501"]
target-version = "py310"

[tool.ruff.isort]
lines-after-imports = 2

# Black configuration
[tool.black]
line-length = 88

[tool.setuptools.packages.find]
where = ["."]
include = ["src"]
exclude = ["tests"]
namespaces = false